new Javaclient created, JavaClientGui is updated to handle that client, main JavaClient also updated, some updates to QueueServer
QueueServer: can now publish the supervisor message. (new method added)
JavaClient:  new method created placeInQueue() method call placeInQueue was moved here, this was done to distinguish between the student and supervisor clients entering seperate queues.

an if statement inside the getCurrentSupervisors() method was added to handle supervisor messages.

the main method was moved to a separate file. JavaClientExecutor

JavaClientGui: was updated to handle the supervisor messages
The gui now has different functionalities depending on if it is opened as a supervisor or student